echo: move the registry type atom into internal/Registry/atoms
Follow the convention of the other atoms in the package: the family lives in
`internal/<Area>/atoms.ts` and the namespace module re-exports the maker.

Flatten the nested family to a structurally-equal `[registry, typename]` tuple
key, matching `annotationPropertyFamily`. Nesting was unsound here: the outer
family memoizes through a `WeakRef`, which is fine for an atom a mounted
consumer keeps alive but not for the inner family a nested lookup returns —
collecting it would silently hand the next caller a second atom with its own
subscription.

No `Atom.keepAlive`, unlike the entity families: the registry is the source of
truth so a remount re-resolves from it, and pinning would hold both a `changed`
listener and the registry itself (now the family key) for the lifetime of the
process.
